fix(web): type-checks
