+++ web/src/components/auth/CallbackPage.tsx
@@ -5,7 +5,7 @@
 
 import { useEffect, useState } from "react";
 import { useNavigate, useSearchParams } from "react-router-dom";
-import { useAuth } from "../../contexts/AuthContext";
+import { useAuth } from "../../hooks/useAuth";
 
 export function CallbackPage() {
   const [error, setError] = useState<string | null>(null);
