+++ web/src/components/auth/withAuth.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { ProtectedRoute } from "./ProtectedRoute";
+
+export function withAuth<P extends object>(
+  WrappedComponent: React.ComponentType<P>
+) {
+  return function WithAuthComponent(props: P) {
+    return (
+      <ProtectedRoute>
+        <WrappedComponent {...props} />
+      </ProtectedRoute>
+    );
+  };
+}
