+++ web/src/hooks/useEventSource.ts
@@ -4,7 +4,7 @@
  */
 
 import { useEffect, useRef, useState, useCallback } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from './useAuth';
 
 // Error types for better error handling
 export enum EventSourceErrorType {
